05.05.2020 v.8

+++ lib/data_loader.rb
@@ -6,6 +6,7 @@ require './models/quiz'
 require './models/scorm_package'
 require './models/task'
 require './models/attachment'
+require './lib/attribute'
 
 module Teachbase
   module Bot
@@ -14,18 +15,11 @@ module Teachbase
 
       MAX_RETRIES = 3
       CS_STATES = %i[active archived].freeze
-      SECTION_OBJECTS = [ :materials, :scorm_packages, :quizzes, :tasks ].freeze
-      SECTION_OBJECTS_CUSTOM_PARAMS = { materials: { "type" => :content_type },
-                                        scorm_packages: { "title" => :name } }.freeze
       MAIN_OBJECTS_CUSTOM_PARAMS = { users: { "name" => :first_name },
                                      course_sessions: { "updated_at" => :changed_at } }.freeze
-      SECTION_OBJECT_TYPES = { materials: :material,
-                       scorm_packages: :scorm_package,
-                       quizzes: :quiz,
-                       tasks: :task}
       CONTENT_VIDEO_FORMAT = "mp4".freeze
 
-      attr_reader :user, :authsession
+      attr_reader :appshell
 
       def initialize(appshell)
         raise "'#{appshell}' is not Teachbase::Bot::AppShell" unless appshell.is_a?(Teachbase::Bot::AppShell)
@@ -35,97 +29,86 @@ module Teachbase
         @retries = 0
       end
 
-      def get_cs_info(cs_id)
-        get_data { user.course_sessions.find_by(tb_id: cs_id) }
-      end
-
-      def get_user_profile
-        get_data { user.profile }
-      end
-
-      def get_cs_list(state, limit_count, offset_num)
-        get_data do
-          user.course_sessions.order(name: :asc)
-                              .limit(limit_count)
-                              .offset(offset_num)
-                              .where(complete_status: state.to_s, scenario_mode: @appshell.settings.scenario)
-        end
-      end
-
-      def get_cs_sections(cs_id)
-        get_data { get_cs_info(cs_id).sections.order(position: :asc) }
+      def get_cs_list(params)
+        state = params[:state]
+        limit_count = params[:limit]
+        offset_num = params[:offset]
+        appshell.user.course_sessions.order(name: :asc)
+                                     .limit(limit_count)
+                                     .offset(offset_num)
+                                     .where(complete_status: state.to_s,
+                                            scenario_mode: appshell.settings.scenario)
       end
 
       def get_cs_sec_by(option, param, cs_id)
-        get_data do
-          case option
-          when :position
-            user.course_sessions.find_by(tb_id: cs_id).sections.find_by(position: param)
-          when :id
-            user.course_sessions.find_by(tb_id: cs_id).sections.find_by(id: param)
-          end
-        end
+        raise "No such option: '#{option}" unless [:position, :id].include?(option.to_sym)
+
+        appshell.user.course_sessions.find_by(tb_id: cs_id).sections.find_by(option.to_sym => param)
       end
 
       def get_cs_sec_contents(section_bd)
         return unless section_bd
 
         section_objects = {}
-        get_data do
-          SECTION_OBJECTS.each do |content_type|
-            section_objects[content_type] = section_bd.public_send(content_type).order(position: :asc)
-          end
+        Teachbase::Bot::Section::OBJECTS.each do |content_type|
+          section_objects[content_type] = section_bd.public_send(content_type).order(position: :asc)
         end
         section_objects
       end
 
       def get_cs_sec_content(content_type, cs_tb_id, sec_id, content_tb_id)
-        get_data do
-          user.course_sessions.find_by(tb_id: cs_tb_id)
-              .sections.find_by(id: sec_id)
-              .public_send(content_type).find_by(tb_id: content_tb_id)
-        end
+          appshell.user.course_sessions.find_by(tb_id: cs_tb_id)
+                  .sections.find_by(id: sec_id)
+                  .public_send(content_type).find_by(tb_id: content_tb_id)
       end
 
       def call_profile
         call_data do
-          lms_info = authsession.load_profile
+          lms_info = appshell.authsession.load_profile
           raise "Profile is not loaded" unless lms_info
 
-          profile = Teachbase::Bot::Profile.find_or_create_by!(user_id: user.id)
-          user_params = create_attributes(Teachbase::Bot::User.attribute_names, lms_info, MAIN_OBJECTS_CUSTOM_PARAMS[:users])
-          user_params[:tb_id] = lms_info["id"]
-          profile_params = create_attributes(Teachbase::Bot::Profile.attribute_names, lms_info)
-          user.update!(user_params)
-          profile.update!(profile_params)
+          user_attrs = Attribute.create(Teachbase::Bot::User.attribute_names, lms_info,
+                                        MAIN_OBJECTS_CUSTOM_PARAMS[:users])
+          user_attrs[:tb_id] = lms_info["id"]
+          profile_attrs = Attribute.create(Teachbase::Bot::Profile.attribute_names, lms_info)
+          @logger.debug "user_attrs: #{user_attrs}\nappshell.user: #{appshell.user}"
+          appshell.user.update!(user_attrs)
+          Teachbase::Bot::Profile.find_or_create_by!(user_id: appshell.user.id).update!(profile_attrs)
         end
       end
 
-      def call_cs_list(state, mode = :normal)
+      def call_cs_list(params)
+        state = params[:state]
+        limit_count = params[:limit]
+        offset_num = params[:offset]
+        mode = params[:mode] || :normal
         raise "No such option for update course sessions list" unless CS_STATES.include?(state.to_sym)
 
         call_data do
           delete_course_sessions(state) if mode == :with_reload
-          lms_info = authsession.load_course_sessions(state)
-          lms_info.each do |course_session_lms|
-            cs = user.course_sessions.find_or_create_by!(tb_id: course_session_lms["id"])
-            course_session_params = create_attributes(Teachbase::Bot::CourseSession.attribute_names,
-                                                      course_session_lms,
-                                                      MAIN_OBJECTS_CUSTOM_PARAMS[:course_sessions])
-            course_session_params[:complete_status] = state.to_s
-            cs.update!(course_session_params)
+          lms_info = appshell.authsession.load_course_sessions(state)
+          ind = offset_num ? offset_num : 0
+          stop_ind = limit_count ? offset_num + limit_count : lms_info.size - 1
+          loop do
+            course_session_lms = lms_info[ind]
+            cs = appshell.user.course_sessions.find_or_create_by!(tb_id: course_session_lms["id"])
+            cs_attrs = Attribute.create(Teachbase::Bot::CourseSession.attribute_names, course_session_lms,
+                                        MAIN_OBJECTS_CUSTOM_PARAMS[:course_sessions])
+            cs_attrs[:complete_status] = state.to_s
+            cs.update!(cs_attrs)
+            ind += 1
+            break if ind == stop_ind + 1 || lms_info[ind].nil?
           end
         end
       end
 
       def call_cs_info(cs_id)
         call_data do
-          cs = user.course_sessions.find_by!(tb_id: cs_id)
-          lms_info = authsession.load_cs_info(cs_id)
-          course_session_params = create_attributes(Teachbase::Bot::CourseSession.attribute_names,
-                                                    lms_info,
-                                                    MAIN_OBJECTS_CUSTOM_PARAMS[:course_sessions])
-          cs.update!(course_session_params)
+          cs = appshell.user.course_sessions.find_by!(tb_id: cs_id)
+          lms_info = appshell.authsession.load_cs_info(cs_id)
+          cs_attrs = Attribute.create(Teachbase::Bot::CourseSession.attribute_names, lms_info,
+                                       MAIN_OBJECTS_CUSTOM_PARAMS[:course_sessions])
+          cs.update!(cs_attrs)
           cs
         end
       end
@@ -134,21 +117,23 @@ module Teachbase
         call_data do
           return if course_session_last_version?(cs_id) && !course_session_last_version?(cs_id).sections.empty?
 
-          cs = user.course_sessions.find_by!(tb_id: cs_id)
+          cs = appshell.user.course_sessions.find_by!(tb_id: cs_id)
           cs.sections.destroy_all if cs
-          authsession.load_cs_info(cs_id)["sections"].each_with_index do |section_lms, ind|
-            section_bd = cs.sections.find_or_create_by!(position: ind + 1, user_id: user.id)
-            section_params = create_attributes(Teachbase::Bot::Section.attribute_names, section_lms)
-            section_bd.update!(section_params)
+          appshell.authsession.load_cs_info(cs_id)["sections"].each_with_index do |section_lms, ind|
+            section_bd = cs.sections.find_or_create_by!(position: ind + 1, user_id: appshell.user.id)
+            section_attrs = Attribute.create(Teachbase::Bot::Section.attribute_names, section_lms)
+            section_bd.update!(section_attrs)
           end
         end
       end
 
-      def call_cs_sec_contents(cs_tb_id, sec_position)
+      def call_cs_sec_contents(section_bd)
+        raise unless section_bd.is_a?(Teachbase::Bot::Section)
+
         call_data do
-          section_bd = get_cs_sec_by(:position, sec_position, cs_tb_id)
-          section_lms = authsession.load_cs_info(cs_tb_id)["sections"][sec_position.to_i - 1]
-          SECTION_OBJECTS.each do |type|
+          cs_data_lms = appshell.authsession.load_cs_info(section_bd.course_session.tb_id)
+          section_lms = cs_data_lms["sections"][section_bd.position - 1]
+          Teachbase::Bot::Section::OBJECTS.each do |type|
             section_bd.public_send(type).destroy_all
             next if section_lms[type.to_s].empty?
             fetch_section_objects(type, section_lms, section_bd)
@@ -161,30 +146,26 @@ module Teachbase
         raise unless section_bd || section_bd.empty?
 
         call_data do
-          lms_data = authsession.load_content(content_type, cs_tb_id, content_tb_id)
+          lms_data = appshell.authsession.load_content(content_type, cs_tb_id, content_tb_id)
+          if lms_data["type"] == "video" && lms_data["source"].is_a?(Hash)
+            lms_data["source"] = lms_data["source"][CONTENT_VIDEO_FORMAT]
+          end
           @logger.debug "lms_data: #{lms_data}"
-          attributes = fetch_content_attr_by_type(content_type, lms_data)
+          attributes = Attribute.create(content_attrs(content_type), lms_data)
           section_bd.public_send(content_type).find_by!(tb_id: content_tb_id).update!(attributes)
         end
       end
 
       def delete_course_sessions(state)
-        get_data { user.course_sessions.where(complete_status: state.to_s).destroy_all }
+        appshell.user.course_sessions.where(complete_status: state.to_s).destroy_all
       end
 
       private
 
-      def get_data
-        @authsession = @appshell.authorizer.call_authsession(:without_api)
-        @user = authsession.user
-        yield
-      end
-
       def call_data
-        return unless @appshell.access_mode == :with_api
+        return unless appshell.access_mode == :with_api
 
-        @authsession = @appshell.authorizer.call_authsession(:with_api)
-        @user = authsession.user
+        appshell.user
         yield
       rescue RuntimeError => e
         if e.http_code == 401 || e.http_code == 403
@@ -199,84 +180,29 @@ module Teachbase
       end
 
       def course_session_last_version?(cs_id)
-        user.course_sessions.find_by(tb_id: cs_id, changed_at: call_cs_info(cs_id).changed_at)
+        appshell.user.course_sessions.find_by(tb_id: cs_id, changed_at: call_cs_info(cs_id).changed_at)
       end
 
-      def fetch_content_attr_by_type(content_type, lms_data)
-        case content_type.to_sym
-        when :materials
-          fetch_content_material(lms_data)
-        when :tasks
-          fetch_content_task(lms_data)
-        when :quizzes
-          #fetch_content_quiz(lms_data)
-        when :scorm_packages
-          #fetch_content_scorm_package(lms_data)
-        else
-          raise "Can't open such content type: '#{content_type}'"
+      def fetch_section_objects(content_type, section_lms, section_bd)
+        raise "No such content type: #{content_type}." unless section_bd.respond_to?(content_type)
+
+        content_params = content_attrs(content_type)
+        section_lms[content_type.to_s].each do |content_type_hash|
+          attributes = Attribute.create(content_params, content_type_hash,
+                                        Teachbase::Bot::Section::OBJECTS_CUSTOM_PARAMS[content_type])
+          section_bd.public_send(content_type).find_or_create_by!(position: content_type_hash["position"],
+                                              tb_id: content_type_hash["id"],
+                                              user_id: appshell.user.id,
+                                              course_session_id: section_bd.course_session.id)
+                                             .update!(attributes)
         end
       end
 
-      def fetch_content_material(material_lms)
-        attributes = {}
-        case material_lms["type"]
-        when "video"
-          attributes[:source] = material_lms["source"][CONTENT_VIDEO_FORMAT]
-        when "vimeo", "youtube", "image", "audio", "pdf", "iframe"
-          attributes[:source] = material_lms["source"]
-        when "text"
-          if material_lms["editor_js"]
-            attributes[:editor_js] = material_lms["editor_js"]
-            attributes[:content] = material_lms["content"]
-          elsif material_lms["markdown"]
-            attributes[:markdown] = material_lms["markdown"]
-            attributes[:content] = material_lms["content"]
-          else
-            attributes[:content] = material_lms["source"]
-          end
-        else 
-          raise "Can't fetch such material type: '#{material_lms["type"]}'"
-        end
-        attributes
+      def content_attrs(content_type)
+         to_constantize(to_camelize(Teachbase::Bot::Section::OBJECTS_TYPES[content_type.to_sym]),
+                        "Teachbase::Bot::").public_send(:attribute_names)
       end
 
-      def fetch_content_task(task_lms)
-        attributes = {}
-      end
-
-      def create_attributes(params, object_type_hash, custom_params = {})
-        raise "Params must be an Array. Your params: #{params} is #{params.class}." unless params.is_a?(Array)
-
-        attributes = {}
-        replace_key_names(custom_params, object_type_hash) if custom_params && !custom_params.empty?
-        params.each do |param|
-          if !object_type_hash[param.to_s].nil? && ![:id, :position].include?(param.to_sym)
-            attributes[param.to_sym] = object_type_hash[param.to_s]
-          end
-        end
-        attributes
-      end
-
-      def fetch_section_objects(conten_type, section_lms, section_bd)
-        raise "No such content type: #{conten_type}." unless section_bd.respond_to?(conten_type)
-
-        content_params = to_constantize(to_camelize(SECTION_OBJECT_TYPES[conten_type]), "Teachbase::Bot::")
-                         .public_send(:attribute_names)
-        cs_id = section_bd.course_session.id        
-        section_lms[conten_type.to_s].each do |content_type_hash|
-          attributes = create_attributes(content_params, content_type_hash, SECTION_OBJECTS_CUSTOM_PARAMS[conten_type])
-          section_bd.public_send(conten_type)
-                    .find_or_create_by!(position: content_type_hash["position"],
-                                        tb_id: content_type_hash["id"],
-                                        user_id: user.id,
-                                        course_session_id: cs_id)
-                    .update!(attributes)
-        end
-      end
-
-      def replace_key_names(mapping, initial_hash)
-        mapping.each { |old_key, new_key| initial_hash[new_key.to_s] = initial_hash.delete(old_key) }
-      end
     end
   end
 end
